requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
libcst==0.4.9
markdown-it-py==2.2.0
mdurl==0.1.2
mypy-extensions==1.0.0
Pygments==2.14.0
pyodbc==4.0.35
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
PyYAML==6.0
rich==13.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
strawberry-graphql==0.168.2
typing-inspect==0.8.0
typing_extensions==4.5.0
uvicorn==0.21.1
wfastcgi==3.0.0
